D2/documentation_backend/requirements.txt
requests==2.31.0  # Для работы с GitHub API
python-dotenv==1.0.0  # Для загрузки переменных окружения (токен GitHub)
anyio==4.9.0
asgiref==3.8.1
asttokens==2.4.0
attrs==25.3.0
autolab-core==1.1.1
backcall==0.2.0
Brotli==1.1.0
cairocffi==1.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
colorlog==6.9.0
comm==0.1.4
contourpy==1.1.1
cssselect2==0.8.0
cycler==0.12.0
debugpy==1.8.0
decorator==5.1.1
dill==0.3.9
Django==5.1.8
django-allauth==65.6.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-yasg==1.21.10
executing==2.0.0
filemagic==1.6
fonttools==4.43.0
freetype-py==2.5.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
inflection==0.5.1
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
lazy_loader==0.4
lxml==5.3.1
manifold3d==3.0.1
mapbox_earcut==1.0.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
multiprocess==0.70.17
nest-asyncio==1.5.8
networkx==3.4.2
numpy==1.26.0
opencv-python==4.11.0.86
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
pillow==11.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycollada==0.9
pycparser==2.22
pydyf==0.11.0
pyglet==2.1.3
Pygments==2.16.1
PyOpenGL==3.1.0
pyparsing==3.1.1
pyphen==0.17.2
pyreadline==2.1
pyrender==0.1.45
python-dateutil==2.8.2
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2025.2
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.1
referencing==0.36.2
rpds-py==0.24.0
rtree==1.4.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
setproctitle==1.3.5
shapely==2.0.7
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
svg.path==6.3
threadpoolctl==3.6.0
tifffile==2025.3.30
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.3.3
traitlets==5.11.2
trimesh==4.6.6
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
vhacdx==0.0.8.post2
visualization==1.0.0
wcwidth==0.2.8
weasyprint==65.0
webencodings==0.5.1
xxhash==3.5.0
zopfli==0.2.3.post1
GitPython
markdown2